Refactor: 1. Move all repository dependency to RepositoryModule 2. Move all mapper dependency to MapperModule
