+++ app/src/main/java/id/novian/flowablecash/di/DatabaseModule.kt
@@ -8,10 +8,6 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import id.novian.flowablecash.data.local.database.AppDatabase
-import id.novian.flowablecash.data.local.repository.BalanceSheetLocalRepository
-import id.novian.flowablecash.data.local.repository.BalanceSheetLocalRepositoryImpl
-import id.novian.flowablecash.data.local.repository.TransactionLocalRepository
-import id.novian.flowablecash.data.local.repository.TransactionLocalRepositoryImpl
 import javax.inject.Singleton
 
 @Module
@@ -28,15 +24,5 @@ object DatabaseModule {
         )
             .fallbackToDestructiveMigration()
             .build()
-    }
-
-    @Provides
-    fun provideLocalRepo(database: AppDatabase): TransactionLocalRepository {
-        return TransactionLocalRepositoryImpl(database.dao())
-    }
-
-    @Provides
-    fun provideBalanceLocalRepository(database: AppDatabase): BalanceSheetLocalRepository {
-        return BalanceSheetLocalRepositoryImpl(database.balanceSheetDao())
     }
 }
